feat(builder): structure builder models into core and versioned v0_9 package
- Extract version-agnostic tree runtime (ComponentBuilderNode, ComponentRef, IdAllocator, ComponentTree, flatten_component_tree) into a2ui.builder.core
- Implement version-specific wire models (Action, FunctionCall, envelopes) under a2ui.builder.v0_9
- Relocate basic catalog to a2ui.builder.v0_9.catalogs.basic with explicit versioned imports
- Remove unversioned latest-defaulting facades to ensure zero breaking changes across protocol evolutions
- Add PEP 561 py.typed marker to a2ui_agent
