+++ agent_sdks/python/a2ui_agent/src/a2ui/py.typed
@@ -1,0 +1,1 @@
+# Marker file for PEP 561.
